Added new weapon sprites and gunsounds
Added two medium autocannon turrets as well as their sounds

The Klanesh, a medium autoturret for prolonged long range (800m) engagement with anti shield and slight EMP capability. It could be a variant of the UAF Tona II railgun, lower range, but better rate of fire.

The Klazer, a medium autoturret specialized for closer engagement (700m). It could be an upgraded version of the Ion Pulser (from the vanilla I think). Dropping heavy EMP rounds on a target.